[REFACTOR] FontSize 별도 파일로 분리 #10
