+++ src/components/atoms/buttons/StrokeButton.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import classNames from 'classnames';
 
+import { FontSize, FONT_SIZE_MAPS } from '@/styles/styleConstants';
+
 export enum Variant {
   PRIMARY = 'primary',
 }
@@ -8,24 +10,6 @@ export enum Variant {
 const VARIANT_MAPS = {
   [Variant.PRIMARY]:
     'hover:bg-primary hover:text-white hover:border-transparent active:bg-gray-light active:text-white active:border-transparent ',
-};
-
-export enum FontSize {
-  XS = 'xs',
-  SM = 'sm',
-  BASE = 'base',
-  LG = 'lg',
-  XL = 'xl',
-  '2XL' = '2xl',
-}
-
-const FONT_SIZE_MAPS = {
-  [FontSize.XS]: 'text-xs',
-  [FontSize.SM]: 'text-sm',
-  [FontSize.BASE]: 'text-base',
-  [FontSize.LG]: 'text-lg',
-  [FontSize.XL]: 'text-xl',
-  [FontSize['2XL']]: 'text-2xl',
 };
 
 export interface StrokeButtonPropsType
@@ -52,6 +36,7 @@ const StrokeButton: React.FC<StrokeButtonPropsType> = ({
     'rounded-full',
     disabled ? 'text-gray-light' : 'text-primary',
     'px-5',
+    'py-1',
     'focus:outline-none',
     'shadow-md',
     VARIANT_MAPS[variant],
