+++ src/components/atoms/chips/ToggleChip.tsx
@@ -1,30 +1,14 @@
 import React from 'react';
 import classNames from 'classnames';
 
+import { FontSize, FONT_SIZE_MAPS } from '@/styles/styleConstants';
+
 export enum Variant {
   PRIMARY = 'primary',
 }
 
 const VARIANT_MAPS = {
   [Variant.PRIMARY]: 'text-primary bg-white border-primary hover:bg-primary-dark hover:text-white',
-};
-
-export enum FontSize {
-  XS = 'xs',
-  SM = 'sm',
-  BASE = 'base',
-  LG = 'lg',
-  XL = 'xl',
-  '2XL' = '2xl',
-}
-
-const FONT_SIZE_MAPS = {
-  [FontSize.XS]: 'text-xs',
-  [FontSize.SM]: 'text-sm',
-  [FontSize.BASE]: 'text-base',
-  [FontSize.LG]: 'text-lg',
-  [FontSize.XL]: 'text-xl',
-  [FontSize['2XL']]: 'text-2xl',
 };
 
 export interface SelectionChipPropsType
