+++ src/styles/styleConstants.ts
@@ -1,0 +1,17 @@
+export enum FontSize {
+  XS = 'xs',
+  SM = 'sm',
+  BASE = 'base',
+  LG = 'lg',
+  XL = 'xl',
+  '2XL' = '2xl',
+}
+
+export const FONT_SIZE_MAPS = {
+  [FontSize.XS]: 'text-xs',
+  [FontSize.SM]: 'text-sm',
+  [FontSize.BASE]: 'text-base',
+  [FontSize.LG]: 'text-lg',
+  [FontSize.XL]: 'text-xl',
+  [FontSize['2XL']]: 'text-2xl',
+};
